+++ covalent/_shared_files/exceptions.py
@@ -21,3 +21,7 @@
 
 class MissingLatticeRecordError(Exception):
     pass
+
+
+class TaskRuntimeError(Exception):
+    pass