+++ covalent_dispatcher/_core/execution.py
@@ -33,6 +33,7 @@ from typing import Any, Dict, List, Tuple
 
 from covalent._results_manager import Result
 from covalent._shared_files import logger
+from covalent._shared_files.config import get_config
 from covalent._shared_files.context_managers import active_lattice_manager
 from covalent._shared_files.defaults import (
     electron_dict_prefix,
@@ -45,7 +46,7 @@ from covalent._workflow import DepsBash, DepsCall, DepsPip
 from covalent._workflow.lattice import Lattice
 from covalent._workflow.transport import TransportableObject
 from covalent.executor import _executor_manager
-from covalent.executor.base import AsyncBaseExecutor, wrapper_fn
+from covalent.executor.base import wrapper_fn
 from covalent_ui import result_webhook
 
 from .._db import update, upsert
@@ -57,6 +58,7 @@ from .._db.write_result_to_db import (
 
 app_log = logger.app_log
 log_stack_info = logger.log_stack_info
+debug_mode = get_config("sdk.log_level") == "debug"
 
 
 # This is to be run out-of-process
@@ -288,12 +290,15 @@ async def _run_task(
         executor = _executor_manager.get_executor(short_name)
         executor.from_dict(object_dict)
     except Exception as ex:
-        app_log.debug(f"Exception when trying to instantiate executor: {ex}")
+        tb = "".join(traceback.TracebackException.from_exception(ex).format())
+        app_log.debug("Exception when trying to instantiate executor:")
+        app_log.debug(tb)
+        error_msg = tb if debug_mode else str(ex)
         node_result = generate_node_result(
             node_id=node_id,
             end_time=datetime.now(timezone.utc),
             status=Result.FAILED,
-            error="".join(traceback.TracebackException.from_exception(ex).format()),
+            error=error_msg,
         )
         return node_result
 
@@ -348,29 +353,38 @@ async def _run_task(
                 results_dir=results_dir,
                 node_id=node_id,
             )
-
-            if isinstance(executor, AsyncBaseExecutor):
-                output, stdout, stderr = await execute_callable()
+            output, stdout, stderr, exception_raised = await executor._execute(
+                function=assembled_callable,
+                args=inputs["args"],
+                kwargs=inputs["kwargs"],
+                dispatch_id=dispatch_id,
+                results_dir=results_dir,
+                node_id=node_id,
+            )
+            if exception_raised:
+                status = Result.FAILED
             else:
-                loop = asyncio.get_running_loop()
-                output, stdout, stderr = await loop.run_in_executor(None, execute_callable)
+                status = Result.COMPLETED
 
             node_result = generate_node_result(
                 node_id=node_id,
                 end_time=datetime.now(timezone.utc),
-                status=Result.COMPLETED,
+                status=status,
                 output=output,
                 stdout=stdout,
                 stderr=stderr,
             )
 
     except Exception as ex:
-        app_log.error(f"Exception occurred when running task {node_id}: {ex}")
+        tb = "".join(traceback.TracebackException.from_exception(ex).format())
+        app_log.debug(f"Exception occurred when running task {node_id}:")
+        app_log.debug(tb)
+        error_msg = tb if debug_mode else str(ex)
         node_result = generate_node_result(
             node_id=node_id,
             end_time=datetime.now(timezone.utc),
             status=Result.FAILED,
-            error="".join(traceback.TracebackException.from_exception(ex).format()),
+            error=error_msg,
         )
     app_log.debug(f"Node result: {node_result}")
     return node_result
@@ -432,8 +446,7 @@ async def _handle_failed_node(result_object, node_result, pending_deps, tasks_qu
     node_id = node_result["node_id"]
     result_object._status = Result.FAILED
     result_object._end_time = datetime.now(timezone.utc)
-    result_object._error = f"Node {result_object._get_node_name(node_id)} failed: \n{result_object._get_node_error(node_id)}"
-    app_log.warning("8A: Failed node upsert statement (run_planned_workflow)")
+    app_log.debug("8A: Failed node upsert statement (run_planned_workflow)")
     upsert._lattice_data(result_object)
     await result_webhook.send_update(result_object)
     await tasks_queue.put(-1)
@@ -442,14 +455,14 @@ async def _handle_failed_node(result_object, node_result, pending_deps, tasks_qu
 async def _handle_cancelled_node(result_object, node_result, pending_deps, tasks_queue):
     result_object._status = Result.CANCELLED
     result_object._end_time = datetime.now(timezone.utc)
-    app_log.warning("9: Failed node upsert statement (run_planned_workflow)")
+    app_log.debug("9: Failed node upsert statement (run_planned_workflow)")
     upsert._lattice_data(result_object)
     await result_webhook.send_update(result_object)
     await tasks_queue.put(-1)
 
 
 async def _update_node_result(result_object, node_result, pending_deps, tasks_queue):
-    app_log.warning("Updating node result (run_planned_workflow).")
+    app_log.debug("Updating node result (run_planned_workflow).")
     update._node(result_object, **node_result)
     await result_webhook.send_update(result_object)
 
@@ -633,10 +646,9 @@ async def _run_planned_workflow(result_object: Result) -> Result:
         node_id = await tasks_queue.get()
         app_log.debug(f"Processing node {node_id}")
 
+        # Tasks queue has been poisoned, don't process the graph further
         if node_id < 0:
-            app_log.debug(f"Workflow {result_object.dispatch_id} failed or cancelled.")
-            await asyncio.gather(*task_futures)
-            return result_object
+            break
 
         # Get name of the node for the current task
         node_name = result_object.lattice.transport_graph.get_node_value(node_id, "name")
@@ -725,6 +737,12 @@ async def _run_planned_workflow(result_object: Result) -> Result:
 
     if result_object._status in [Result.FAILED, Result.CANCELLED]:
         app_log.debug(f"Workflow {result_object.dispatch_id} cancelled or failed")
+
+        failed_nodes = result_object._get_failed_nodes()
+        failed_nodes = map(lambda x: f"{x[0]}: {x[1]}", failed_nodes)
+        failed_nodes_msg = "\n".join(failed_nodes)
+        result_object._error = "The following tasks failed:\n" + failed_nodes_msg
+        upsert._lattice_data(result_object)
         return result_object
 
     app_log.debug("8: All tasks finished running (run_planned_workflow)")
@@ -792,6 +810,11 @@ async def run_workflow(result_object: Result) -> Result:
             updated_at=datetime.now(timezone.utc),
         )
 
+        error_msg = "".join(traceback.TracebackException.from_exception(ex).format())
+        result_object._status = Result.FAILED
+        result_object._error = error_msg
+        result_object._end_time = datetime.now(timezone.utc)
+        upsert._lattice_data(result_object)
         write_lattice_error(
             result_object.dispatch_id,
             "".join(traceback.TracebackException.from_exception(ex).format()),