+++ tests/covalent_tests/executor/executor_plugins/local_test.py
@@ -20,12 +20,17 @@
 
 """Tests for Covalent local executor."""
 
+import io
 import tempfile
 from functools import partial
 
+import pytest
+
 import covalent as ct
+from covalent._shared_files import TaskRuntimeError
 from covalent._workflow.transport import TransportableObject
-from covalent.executor.executor_plugins.local import LocalExecutor, wrapper_fn
+from covalent.executor.base import wrapper_fn
+from covalent.executor.executor_plugins.local import LocalExecutor
 
 
 def test_local_executor_passes_results_dir(mocker):
@@ -115,3 +120,42 @@ def test_local_executor_run():
     kwargs = {}
     task_metadata = {"dispatch_id": "asdf", "node_id": 1}
     assert le.run(f, args, kwargs, task_metadata) == 25
+
+
+def test_local_executor_run_exception_handling(mocker):
+    def f(x):
+        print("f output")
+        raise RuntimeError("error")
+
+    le = LocalExecutor()
+    le._task_stdout = io.StringIO()
+    le._task_stderr = io.StringIO()
+    args = [5]
+    kwargs = {}
+    task_metadata = {"dispatch_id": "asdf", "node_id": 1}
+    with pytest.raises(TaskRuntimeError) as ex:
+        le.run(f, args, kwargs, task_metadata)
+    le._task_stdout.getvalue() == "f output"
+    assert "RuntimeError" in le._task_stderr.getvalue()
+
+
+def test_local_wrapper_fn_exception_handling(mocker):
+    import multiprocessing as mp
+
+    from covalent.executor.utils.wrappers import local_wrapper
+
+    def failing_task():
+        raise RuntimeError("Err")
+
+    args = [5]
+    kwargs = {}
+    error_msg = "task failed"
+    q = mp.Queue()
+    mocker.patch("traceback.TracebackException.from_exception", return_value=error_msg)
+    p = mp.Process(target=local_wrapper, args=(failing_task, args, kwargs, q))
+    p.start()
+    p.join()
+    output, stdout, stderr, tb = q.get(False)
+
+    assert tb == error_msg
+    assert output is None